requirements.txt
google-adk>=1.15.0,<2.0.0
opentelemetry-instrumentation-google-genai>=0.1b0,<1.0.0
gcsfs>=2024.11.0
google-cloud-logging>=3.12.0,<4.0.0
google-cloud-aiplatform[evaluation,agent-engines]>=1.130.0
protobuf>=6.31.1,<7.0.0
python-docx==1.2.0
docxtpl[subdoc]==0.20.2
diagrams==0.25.1